Make Profile Main Expanded View, Profile Education Expanded View and add navigation
This commit creates Profile Main Expanded View and Profile Education Expanded View in a very less detailed manner.
It also adds the navigation from the education card to the Education Expanded View.
There was a small bug which caused the profile view being unable to load, that has been resolved as well.